pyproject.toml
[project]
name = "fortnite-collect-bot"
version = "0.1.0"
description = "Telegram-бот для сбора скуада в Fortnite"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiogram>=3.20,<4",
    "python-dotenv>=1.1",
    "aiosqlite>=0.21",
    "openai>=1.60",
]
